[wip] wire PyVersionedObjects into the database_connection.
Still some broken tests. And We'll need to get all of the view under
C++ for it to actually be fast because there's so much overhead
in the piping between 'object' and the actual view.

I estimate looking up the top object can be done 10mm/sec on my box.
I get about 200k lookups / sec total. That difference is mostly from
the many python instructions we have to execute between the object
__getattr__ call and actually requesting a particular field from
the data store.